+++ NEWS.md
@@ -1,4 +1,24 @@
-## rPHG 0.5
+## rPHG2 0.6
+* Added new parameters to `plogtGvcf()`:
+  + `mData` and `mVar`
+  + allows user to override default sample color option with categorical data
+* Added new generic, `numberOfHaplotypes()`:
+  + returns the number of unique haplotypes found within each reference range or
+    the total value found within a `PHGDataSet` object
+* Added getters to `PHGDataSet` objects:
+  + `numberOfChromosomes()`
+  + `numberOfSamples()`
+  + `numberOfRefRanges()`
+* Changed functions and variables containing `taxa` to `samples`
+  + e.g., `numberOfTaxa()` is now `numberOfSamples()`
+* Format tweaks to `HaplotypeGraph` objects' `show()` method
+* Added new function `plotHaploCount()`:
+  + Plots unique haplotype counts for all or given reference ranges
+* Added new function `plotHaploDist()`:
+  + Plots unique haplotype distribution for entire `PHGDataSet` object
+
+
+## rPHG2 0.5
 * Added new visualization method, `plotGvcf()`:
   + auto plotting various gVCF metrics
   + granular metric support through formula subsetting